Explicitly renaming columns in metadata_csv_filter
This script was originally limiting the input columns on the read_csv() call, but this failed when trying to refilter a file since the input columns have already been renamed to the output columns.  Now the script attempts to rename the field that are needed as the very first operation, and then all extra fields are skipped/ignored.

Also, removed lots of old/unused code.