Standardized script inputs and input parsing
wrs2, year, and month parameters should now all be able to take comma and/or space separated lists and ranges of values.  Switched all the documentation to use space separated values as the default and example, since that seems to be the more standard way of handling multiple inputs to argparse.