Better handling for processing small year ranges and other bug fixes
Standardized and added "years" command line arguments for all scripts.  The scripts will now skip CSV files if they don't include any data for the target year range.  We should probably encourage the users to use the "years" parameters for all the scripts, since it helps avoid processing files that aren't needed.

The one change we may want to add is to remove a CSV file in the filter step if the years don't intersect.  I think this would be the right approach since the workflow is to run each script once, not iteratively.

Closes #18, Closes #17 , Closes #16